corrected the file names
