+++ ascent-platform-starters/ascent-platform-starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,5 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-gov.va.ascent.starter.security.autoconfigure.AscentServiceAutoConfiguration,\
-gov.va.ascent.starter.security.autoconfigure.AscentRestAutoConfiguration,\
-gov.va.ascent.starter.security.autoconfigure.AscentAuditAutoConfiguration
-
+gov.va.ascent.starter.service.autoconfigure.AscentServiceAutoConfiguration,\
+gov.va.ascent.starter.rest.autoconfigure.AscentRestAutoConfiguration,\
+gov.va.ascent.starter.audit.autoconfigure.AscentAuditAutoConfiguration
