+++ ascent-platform-starters/ascent-platform-starter/src/test/java/gov/va/ascent/starter/audit/autoconfigure/AscentAuditAutoConfigurationTest.java
@@ -1,10 +1,7 @@
-package gov.va.ascent.starter.rest.autoconfigure;
+package gov.va.ascent.starter.audit.autoconfigure;
 
 import org.junit.After;
 import org.junit.Test;
-import org.springframework.boot.autoconfigure.security.SecurityAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.ServerPropertiesAutoConfiguration;
-import org.springframework.security.web.FilterChainProxy;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
 import static org.junit.Assert.*;
@@ -12,7 +9,7 @@ import static org.junit.Assert.*;
 /**
  * Created by rthota on 8/24/17.
  */
-public class AscentServiceAutoConfigurationTest {
+public class AscentAuditAutoConfigurationTest {
 
     private AnnotationConfigWebApplicationContext context;
 
@@ -26,10 +23,9 @@ public class AscentServiceAutoConfigurationTest {
 /*    @Test
     public void testWebConfiguration() throws Exception {
         context = new AnnotationConfigWebApplicationContext();
-        context.register(SecurityAutoConfiguration.class, ServerPropertiesAutoConfiguration.class, AscentAuditAutoConfiguration.class);
+        context.register(AscentAuditAutoConfiguration.class);
         context.refresh();
         assertNotNull(context);
-        assertEquals(this.context.getBean(FilterChainProxy.class).getFilterChains().size(), 11);
-
+        assertNotNull(this.context.getBean(AscentAuditAutoConfiguration.class));
     }*/
 }
