+++ ascent-platform-starters/ascent-platform-starter/src/test/java/gov/va/ascent/starter/rest/autoconfigure/AscentRestAutoConfigurationTest.java
@@ -1,0 +1,31 @@
+package gov.va.ascent.starter.rest.autoconfigure;
+
+import org.junit.After;
+import org.junit.Test;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by rthota on 8/24/17.
+ */
+public class AscentRestAutoConfigurationTest {
+
+    private AnnotationConfigWebApplicationContext context;
+
+    @After
+    public void close() {
+        if (this.context != null) {
+            this.context.close();
+        }
+    }
+
+    @Test
+    public void testWebConfiguration() throws Exception {
+        context = new AnnotationConfigWebApplicationContext();
+        context.register(AscentRestAutoConfiguration.class);
+        context.refresh();
+        assertNotNull(context);
+        assertNotNull(this.context.getBean(AscentRestAutoConfiguration.class));
+    }
+}
